+++ src/routes/venues/proposal/+page.ts
@@ -3,6 +3,7 @@ import type { PageLoad } from './$types.js';
 export const load: PageLoad = async ({ parent, params }) => {
 	const { supabase } = await parent();
 
+	// Get all currencies, so the page can show them in a dropdown.
 	const { data: currencies, error: currenciesError } = await supabase
 		.from('currencies')
 		.select('*');